Closure#arity
